+++ src/screens/ManageTag/EditUserTagModal.tsx
@@ -2,7 +2,7 @@ import type { TFunction } from 'i18next';
 import type { FormEvent } from 'react';
 import React from 'react';
 import { Button, Form, Modal } from 'react-bootstrap';
-import styles from '../../style/app.module.css';
+import styles from '../../style/app-fixed.module.css';
 
 /**
  * Edit UserTag Modal component for the Manage Tag screen.
