+++ src/screens/Requests/Requests.tsx
@@ -11,7 +11,7 @@ import TableLoader from 'components/TableLoader/TableLoader';
 import RequestsTableItem from 'components/RequestsTableItem/RequestsTableItem';
 import InfiniteScroll from 'react-infinite-scroll-component';
 import type { InterfaceQueryMembershipRequestsListItem } from 'utils/interfaces';
-import styles from '../../style/app.module.css';
+import styles from '../../style/app-fixed.module.css';
 import useLocalStorage from 'utils/useLocalstorage';
 import { useParams } from 'react-router-dom';
 import SearchBar from 'subComponents/SearchBar';
